build: 移除nohup.out文件

+++ start.sh
@@ -4,7 +4,10 @@
 path=$(cd `dirname "$0"`;pwd)
 
 # 移除 nohup.out 文件
-$(rm $path/nohup.out)
+if [ -e $path/nohup.out ]; then
+  $(rm $path/nohup.out)
+  echo "remove $path/nohup.out"
+fi
 
 # 启动服务
 $(nohup jekyll serve --config $path/_config.yml -s $path/ -d $path/_site > $path/nohup.out &)
